requirements.txt
asn1crypto==0.22.0
certifi==2017.7.27.1
cffi==1.11.0
chardet==3.0.4
cryptography==2.0.3
Django==1.11.5
django-cors-headers==2.1.0
djangorestframework==3.6.4
enum34==1.1.6
idna==2.6
ipaddress==1.0.18
pycparser==2.18
PyJWT==1.5.3
pyOpenSSL==17.3.0
pytz==2017.2
requests==2.18.4
six==1.11.0
twilio==6.6.3
urllib3==1.22
